fix some names, fix lints, move to separate folders and update imports
