+++ crates/testing/src/predicates/event.rs
@@ -1,49 +1,20 @@
-use std::sync::Arc;
-
 use async_trait::async_trait;
-use hotshot::types::SystemContextHandle;
-use hotshot_example_types::node_types::{MemoryImpl, TestTypes};
-use hotshot_task_impls::{
-    consensus::ConsensusTaskState,
-    events::{HotShotEvent, HotShotEvent::*},
-};
+use hotshot_task_impls::events::{HotShotEvent, HotShotEvent::*};
 use hotshot_types::{
     data::null_block,
-    simple_certificate::UpgradeCertificate,
     traits::{block_contents::BlockHeader, node_implementation::NodeType},
 };
+use std::sync::Arc;
 
-#[derive(Eq, PartialEq, Copy, Clone, Debug)]
-pub enum PredicateResult {
-    Pass,
+use crate::predicates::{Predicate, PredicateResult};
 
-    Fail,
-
-    Incomplete,
-}
-
-impl From<bool> for PredicateResult {
-    fn from(boolean: bool) -> Self {
-        match boolean {
-            true => PredicateResult::Pass,
-            false => PredicateResult::Fail,
-        }
-    }
-}
-
-#[async_trait]
-pub trait Predicate<INPUT>: std::fmt::Debug {
-    async fn evaluate(&self, input: &INPUT) -> PredicateResult;
-    async fn info(&self) -> String;
-}
-
-type AsyncBooleanCallback<TYPES> = Arc<dyn Fn(Arc<HotShotEvent<TYPES>>) -> bool + Send + Sync>;
+type BooleanCallback<TYPES> = Arc<dyn Fn(Arc<HotShotEvent<TYPES>>) -> bool + Send + Sync>;
 
 pub struct EventBooleanPredicate<TYPES>
 where
     TYPES: NodeType + Send + Sync,
 {
-    check: AsyncBooleanCallback<TYPES>,
+    check: BooleanCallback<TYPES>,
     info: String,
 }
 
@@ -74,7 +45,7 @@ where
     let info = format!("{:?}", event);
     let event = Arc::new(event);
 
-    let check: AsyncBooleanCallback<TYPES> = Arc::new(move |e: Arc<HotShotEvent<TYPES>>| {
+    let check: BooleanCallback<TYPES> = Arc::new(move |e: Arc<HotShotEvent<TYPES>>| {
         let event_clone = event.clone();
         *e == *event_clone
     });
@@ -87,7 +58,7 @@ where
     TYPES: NodeType,
 {
     let info = "LeafDecided".to_string();
-    let check: AsyncBooleanCallback<TYPES> =
+    let check: BooleanCallback<TYPES> =
         Arc::new(move |e: Arc<HotShotEvent<TYPES>>| matches!(e.as_ref(), LeafDecided(_)));
 
     Box::new(EventBooleanPredicate { check, info })
@@ -98,7 +69,7 @@ where
     TYPES: NodeType,
 {
     let info = "QuorumVoteSend".to_string();
-    let check: AsyncBooleanCallback<TYPES> =
+    let check: BooleanCallback<TYPES> =
         Arc::new(move |e: Arc<HotShotEvent<TYPES>>| matches!(e.as_ref(), QuorumVoteSend(_)));
 
     Box::new(EventBooleanPredicate { check, info })
@@ -109,7 +80,7 @@ where
     TYPES: NodeType,
 {
     let info = "ViewChange".to_string();
-    let check: AsyncBooleanCallback<TYPES> =
+    let check: BooleanCallback<TYPES> =
         Arc::new(move |e: Arc<HotShotEvent<TYPES>>| matches!(e.as_ref(), ViewChange(_)));
     Box::new(EventBooleanPredicate { check, info })
 }
@@ -119,7 +90,7 @@ where
     TYPES: NodeType,
 {
     let info = "UpgradeCertificateFormed".to_string();
-    let check: AsyncBooleanCallback<TYPES> = Arc::new(move |e: Arc<HotShotEvent<TYPES>>| {
+    let check: BooleanCallback<TYPES> = Arc::new(move |e: Arc<HotShotEvent<TYPES>>| {
         matches!(e.as_ref(), UpgradeCertificateFormed(_))
     });
     Box::new(EventBooleanPredicate { check, info })
@@ -130,7 +101,7 @@ where
     TYPES: NodeType,
 {
     let info = "QuorumProposalSend with UpgradeCertificate attached".to_string();
-    let check: AsyncBooleanCallback<TYPES> =
+    let check: BooleanCallback<TYPES> =
         Arc::new(move |e: Arc<HotShotEvent<TYPES>>| match e.as_ref() {
             QuorumProposalSend(proposal, _) => proposal.data.upgrade_certificate.is_some(),
             _ => false,
@@ -143,7 +114,7 @@ where
     TYPES: NodeType,
 {
     let info = "QuorumProposalValidated".to_string();
-    let check: AsyncBooleanCallback<TYPES> = Arc::new(move |e: Arc<HotShotEvent<TYPES>>| {
+    let check: BooleanCallback<TYPES> = Arc::new(move |e: Arc<HotShotEvent<TYPES>>| {
         matches!(*e.clone(), QuorumProposalValidated(..))
     });
     Box::new(EventBooleanPredicate { check, info })
@@ -154,7 +125,7 @@ where
     TYPES: NodeType,
 {
     let info = "QuorumProposalSend".to_string();
-    let check: AsyncBooleanCallback<TYPES> =
+    let check: BooleanCallback<TYPES> =
         Arc::new(move |e: Arc<HotShotEvent<TYPES>>| matches!(e.as_ref(), QuorumProposalSend(..)));
     Box::new(EventBooleanPredicate { check, info })
 }
@@ -166,7 +137,7 @@ where
     TYPES: NodeType,
 {
     let info = "QuorumProposalSend with null block payload".to_string();
-    let check: AsyncBooleanCallback<TYPES> =
+    let check: BooleanCallback<TYPES> =
         Arc::new(move |e: Arc<HotShotEvent<TYPES>>| match e.as_ref() {
             QuorumProposalSend(proposal, _) => {
                 Some(proposal.data.block_header.payload_commitment())
@@ -182,48 +153,7 @@ where
     TYPES: NodeType,
 {
     let info = "TimeoutVoteSend".to_string();
-    let check: AsyncBooleanCallback<TYPES> =
+    let check: BooleanCallback<TYPES> =
         Arc::new(move |e: Arc<HotShotEvent<TYPES>>| matches!(e.as_ref(), TimeoutVoteSend(..)));
     Box::new(EventBooleanPredicate { check, info })
-}
-
-type ConsensusTaskTestState =
-    ConsensusTaskState<TestTypes, MemoryImpl, SystemContextHandle<TestTypes, MemoryImpl>>;
-
-type AsyncUpgradeCertCallback =
-    Arc<dyn Fn(Arc<Option<UpgradeCertificate<TestTypes>>>) -> bool + Send + Sync>;
-
-pub struct UpgradeCertPredicate {
-    check: AsyncUpgradeCertCallback,
-    info: String,
-}
-
-impl std::fmt::Debug for UpgradeCertPredicate {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.info)
-    }
-}
-
-#[async_trait]
-impl Predicate<ConsensusTaskTestState> for UpgradeCertPredicate {
-    async fn evaluate(&self, input: &ConsensusTaskTestState) -> PredicateResult {
-        let upgrade_cert = input.decided_upgrade_cert.clone();
-        PredicateResult::from((self.check)(upgrade_cert.into()))
-    }
-
-    async fn info(&self) -> String {
-        self.info.clone()
-    }
-}
-
-pub fn no_decided_upgrade_cert() -> Box<UpgradeCertPredicate> {
-    let info = "expected decided_upgrade_cert to be None".to_string();
-    let check: AsyncUpgradeCertCallback = Arc::new(move |s| s.is_none());
-    Box::new(UpgradeCertPredicate { info, check })
-}
-
-pub fn decided_upgrade_cert() -> Box<UpgradeCertPredicate> {
-    let info = "expected decided_upgrade_cert to be Some(_)".to_string();
-    let check: AsyncUpgradeCertCallback = Arc::new(move |s| s.is_some());
-    Box::new(UpgradeCertPredicate { info, check })
 }
