+++ crates/testing/tests/tests_1/proposal_ordering.rs
@@ -2,7 +2,7 @@ use hotshot::{tasks::task_state::CreateTaskState, types::SystemContextHandle};
 use hotshot_example_types::node_types::{MemoryImpl, TestTypes};
 use hotshot_task_impls::{consensus::ConsensusTaskState, events::HotShotEvent::*};
 use hotshot_testing::{
-    predicates::{exact, quorum_proposal_send, quorum_proposal_validated},
+    predicates::event::{exact, quorum_proposal_send, quorum_proposal_validated},
     task_helpers::vid_scheme_from_view_number,
     test_helpers::permute_input_with_index_order,
     view_generator::TestViewGenerator,
